Adjusting the network-das code structure
